% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_summary}
\alias{classification_summary}
\title{Summary statistics for the classification data sets}
\format{A data frame with 15 variables:
\describe{
  \item{MajorityClassSize:}{Number of instances in majority class of target variable}
  \item{MinorityClassSize:}{Number of instances in minority class of target variable}
  \item{NumberOfClasses:}{Number of classes in target variable}
  \item{ImbalanceMetric:}{Imbalance metric, where zero means that the dataset is perfectly balanced and the higher the value, the more imbalanced the dataset}
  \item{NumberOfFeatures:}{Total number of features (equal to number of columns)}
  \item{NumberOfBinaryFeatures:}{Number of binary features}
  \item{NumberOfIntegerFeatures:}{Number of integer features}
  \item{NumberOfFloatFeatures:}{Number of float features}
  \item{NumberOfInstances:}{Number of data observations (equal to number of rows)}
  \item{NumberOfInstancesWithMissingValues:}{Number of instances with missing values (always 0)}
  \item{NumberOfMissingValues:}{Number of missing values (always 0)}
  \item{NumberOfNumericFeatures:}{Number of numeric features}
  \item{NumberOfSymbolicFeatures:}{Number of symbolic features}
  \item{name:}{Dataset name}
  \item{status:}{All datasets are currently categorised as 'active'}
}}
\source{
\url{https://github.com/EpistasisLab/penn-ml-benchmarks}
}
\usage{
classification_summary
}
\description{
Summary statistics for the classification data sets
}
\keyword{datasets}
