% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_names}
\alias{dataset_names}
\title{Names of all available datasets}
\format{
An object of class \code{character} of length 286.
}
\source{
\url{https://github.com/EpistasisLab/penn-ml-benchmarks}
}
\usage{
dataset_names
}
\description{
A list of the names of available datasets
}
\keyword{datasets}
