% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regression_summary}
\alias{regression_summary}
\title{Summary statistics for the regression data sets}
\format{A data frame with 13 variables:
\describe{
  \item{did:}{OpenML identifier (see \url{https://www.openml.org/)}}
  \item{MajorityClassSize:}{Number of instances in majority class of target variable (always -1)}
  \item{MaxNominalAttDistinctValues:}{Maximum number of distinct values among attributes of the nominal type}
  \item{MinorityClassSize:}{Number of instances in minority class of target variable (always -1)}
  \item{NumberOfClasses:}{Number of classes in target variable}
  \item{NumberOfFeatures:}{Total number of features (equal to number of columns)}
  \item{NumberOfInstances:}{Number of data observations (equal to number of rows)}
  \item{NumberOfInstancesWithMissingValues:}{Number of instances with missing values (always 0)}
  \item{NumberOfMissingValues:}{Number of missing values (always 0)}
  \item{NumberOfNumericFeatures:}{Number of numeric features}
  \item{NumberOfSymbolicFeatures:}{Number of symbolic features}
  \item{name:}{Dataset name}
  \item{status:}{All datasets are currently categorised as 'active'}
}

For further details, see \url{https://www.openml.org/}}
\source{
\url{https://github.com/EpistasisLab/penn-ml-benchmarks}
}
\usage{
regression_summary
}
\description{
Summary statistics for the regression data sets
}
\keyword{datasets}
